hasher-matcher-actioner/lambdas/requirements.txt
boto
threatexchange[faiss,pdq_hasher]>=0.0.11
